requirements-local.txt
# Core utilities
loguru==0.7.3
networkx>=3.2.1
rich==14.1.0
requests==2.32.4
python-dotenv==1.1.0
prompt-toolkit>=3.0.50

# Config & validation
pydantic>=2.11,<3
pydantic-settings>=2.10.1

# LangChain core
langchain==0.3.27
langchain-community==0.3.27
langchain-ollama==0.3.6

# Embeddings / DL stack
torch>=2.7        # 或按你的平台装指定轮子
transformers>=4.55.0,<5
sentence-transformers>=5.1.0,<6

# Vector DB & search
weaviate-client>=4.16.6
rank-bm25==0.2.2

# Tree-sitter core & grammars
tree-sitter==0.25.1
tree-sitter-python>=0.23.6
tree-sitter-javascript>=0.23.1
tree-sitter-typescript>=0.23.2
tree-sitter-rust>=0.24.0
tree-sitter-go>=0.23.4
tree-sitter-scala>=0.24.0
tree-sitter-java>=0.23.5
tree-sitter-cpp>=0.23.4
